Guess the listening address family in UDP input
This code path used to break for IPv6 addresses with
`SocketError (getaddrinfo: Address family for hostname not supported)`
because it was using `Socket::AF_INET` for every input.

This is related to a JRuby bug[1] that is marked as `wontfix`.

With the new implementation the address family is being retrieved by
`Socket.getaddrinfo` and supports both IPv4 and IPv6 addresses.

[1]: https://github.com/jruby/jruby/issues/5112